chores: extract apigear olink end olink examples
system client and server were split into: an apigear olink package which
contains comon client and server functionality and a examles directory
which uses this functionality along with generated olink adapters.
The examples were given a path context, so all the imports work.
