+++ templates/apigear/olink/__init__.py
@@ -1,0 +1,2 @@
+from .client import Client
+from .remoteendpoint import RemoteEndpoint
